==> Notebooks/01_DS_data_exploration.ipynb <==
import pandas as pd

# Final demo dataset

df = pd.read_csv("../Data/Raw/df_final_demo.csv")
df.head()

len(df)

df.tail()

df.describe()

df["client_id"].duplicated().sum()

df['gendr'].value_counts() 

df.isna().sum()

df[df.isnull().any(axis=1)]


df['num_accts'].min(), df['num_accts'].max()

df['clnt_age'].min(), df['clnt_age'].max()

df['calls_6_mnth'].min(), df['calls_6_mnth'].max()

# Experiment clients dataset

df1 = pd.read_csv("../Data/Raw/df_final_experiment_clients.csv")
df1.head()

len(df1)

df["client_id"].duplicated().sum()

df1["Variation"].value_counts()

df1['Variation'].isna().sum()


# Webdata dataset - Part-1 

df2 = pd.read_csv("../Data/Raw/df_final_web_data_pt_1.csv")
df2.head()

len(df2)

df2["process_step"].unique()

df2["process_step"].value_counts()

df2["client_id"].value_counts()

df2['client_id'].nunique()

steps_per_client = (
    df2.groupby('client_id')['process_step']
    .nunique()
    .sort_values(ascending=False)
)
steps_per_client.describe()

- Median is 5, so at least 50% of the clients completed the process
- Mean is 3.97, denotes that some clients dropped off during the process

# Webdata dataset - Part-2

df3 = pd.read_csv("Data/Raw/df_final_web_data_pt_2.csv")
df3.head()

len(df3)

df3["process_step"].unique()

df3["process_step"].value_counts()

df3["client_id"].value_counts()

df3['client_id'].nunique()

steps_per_client = (
    df3.groupby('client_id')['process_step']
    .nunique()
    .sort_values(ascending=False)
)
steps_per_client.describe()

- Median is 5, so at least 50% of the clients completed the process
- Mean is 4.02, denotes that some clients dropped off during the later stages of the process
