==> Notebooks/02_rq_df_cleaning.ipynb <==
# Final demo dataset

## Loading and exploring the dataframe 

import pandas as pd
import os
import numpy as np

df = pd.read_csv("../Data/Raw/df_final_demo.csv")
df.head(20)

# Checking unique values of each columns
df.nunique()


## Standardising Column names

# Renaming columns with typos and saving cleaned dataframe

df = df.rename(columns={
    "gendr": "gender",
    "num_acts": "num_accts",
    "bal": "clnt_balance"
})


final_demo = "../Data/Processed/df_final_demo_clean.csv"
df.to_csv("../Data/Processed/df_final_demo_clean.csv", index=False)
df = pd.read_csv(final_demo)
df.head()

## Grouping 'gender' values

df["gender"].value_counts()

df[df["gender"] == "X"]


- Since we do not know the gender of these clients, we group them to "U" (Unknown).

df.loc[df["gender"] == "X", "gender"] = "Unknown"

df.loc[df["gender"] == "F", "gender"] = "Female"

df.loc[df["gender"] == "M", "gender"] = "Male"

df.loc[df["gender"] == "U", "gender"] = "Unknown"

## Dropping the null values (only the rows that have all field values as null)

df = df.dropna(
    subset=['clnt_tenure_yr','clnt_age','num_accts','clnt_balance','calls_6_mnth','logons_6_mnth'],
    how='all'
)

# Saving changes again
df.to_csv(final_demo, index=False)

df["gender"].value_counts()

df.nunique()

### Clients Balance column inspection

* We can also see that we have 70609 client id's but only 70327 unique values for client balance, reason to be figured out 

df["clnt_balance"].value_counts()


* It is because some values are shared by multiple clients

# Clients who share the same balance - sanity check
df[df["clnt_balance"] == 31188.62]

df[df["clnt_balance"].duplicated(keep=False)].sort_values("clnt_balance")


df["clnt_balance"].isna().sum()


--------------------------------------------------------------------------------------------------------------------
--------------------------------------------------------------------------------------------------------------------

# Final Experiment Clients Dataset

## Loading and exploring the dataframe 

df_2 = pd.read_csv("../Data/Raw/df_final_experiment_clients.csv")
df_2.head(20)

df_2.isna().sum()

df_2 = df_2.rename(columns={"Variation": "variation"})


df_2["variation"].value_counts(dropna=False)


(df_2["variation"].isna().mean() * 100).round(2)


* 28.5% of the DataFrame consists of NaN values

df_nan = df_2[df_2["variation"].isna()]
df_nan.head(10)

## Replacing NaN values with "Excluded"

# Changing NaN values in to "Excluded" -> This represents clients who weren't a part of the experiment.
df_2["variation"] = df_2["variation"].fillna("excluded")


# saving changes
OUTPUT_DIR = "../Data/Processed"
os.makedirs(OUTPUT_DIR, exist_ok=True)

OUTPUT_FILE = os.path.join(OUTPUT_DIR, "df_final_experiment_clients_clean.csv")
df_2.to_csv(OUTPUT_FILE, index=False)

* More than 25% of our clients weren't part of the test. This is not an error, doesn't need cleaning

-------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------

# Webdata Datasets - pt_1 and pt_2

## Loading the dataframes

df_3_1 = pd.read_csv("../Data/Raw/df_final_web_data_pt_1.csv")
df_3_1.head(5)

df_3_2 = pd.read_csv("../Data/Raw/df_final_web_data_pt_2.csv")
df_3_2.head(5)

## Merging Final Webdata Dataframes into a single dataframe

# Merging 

df_3_1 = pd.read_csv("../Data/Raw/df_final_web_data_pt_1.csv")
df_3_2 = pd.read_csv("../Data/Raw/df_final_web_data_pt_2.csv")

df_final_web = pd.concat([df_3_1, df_3_2], ignore_index=True)

# Sorting by time and client 
df_final_web["date_time"] = pd.to_datetime(df_final_web["date_time"])
df_final_web = df_final_web.sort_values(["client_id", "date_time"]).reset_index(drop=True)

# Saving changes in Processed folder
OUTPUT_FILE = "../Data/Processed/df_final_web_merged.csv"
df_final_web.to_csv(OUTPUT_FILE, index=False)


df_final_web.head(15)

## Separating Date and Time stamps

df_web = pd.read_csv("../Data/Processed/df_final_web_merged.csv")

df_web["date_time"] = pd.to_datetime(df_web["date_time"])

df_web["date"] = df_web["date_time"].dt.date
df_web["time"] = df_web["date_time"].dt.time

# Dropping original column
df_web = df_web.drop(columns=["date_time"])

df_web.head()

# Saving changes

OUTPUT_DIR = "../Data/Processed"
os.makedirs(OUTPUT_DIR, exist_ok=True)

OUTPUT_FILE = os.path.join(OUTPUT_DIR, "df_final_web_merged.csv")
df_web.to_csv(OUTPUT_FILE, index=False)
