==> Notebooks/03_DS_monthly_usage_analysis.ipynb <==
import pandas as pd

#working on a copy
web_merged = pd.read_csv("../Data/Processed/df_final_web_merged.csv")
df = web_merged.copy()

#timestamp creation 
df["date_time"] = pd.to_datetime(
    df["date"].astype(str) + " " + df["time"].astype(str),
    errors="coerce"
)

#calculating active time per visit
visit_level = (
    df.groupby(["client_id", "visit_id"])["date_time"]
      .agg(visit_start="min", visit_end="max")
      .reset_index()
)

visit_level["visit_active_seconds"] = (
    visit_level["visit_end"] - visit_level["visit_start"]
).dt.total_seconds().fillna(0)

#assigning each visit to a month (month start date)
visit_level["month"] = visit_level["visit_start"].dt.to_period("M").dt.to_timestamp()

#aggregateing to client_id x month ---
df_client_usage = (
    visit_level.groupby(["client_id", "month"], as_index=False)
      .agg(
          visits=("visit_id", "nunique"),
          total_active_seconds=("visit_active_seconds", "sum")
      )
)

#adding a month label for Tableau friendly format
df_client_usage["month_label"] = pd.to_datetime(df_client_usage["month"]).dt.strftime("%d-%b-%Y")

#active time in mm:ss
secs = df_client_usage["total_active_seconds"].round().astype("int64")
df_client_usage["total_active_mmss"] = (
    (secs // 60).astype(str).str.zfill(2) + ":" + (secs % 60).astype(str).str.zfill(2)
)

#saving as a new dataset
df_client_usage.to_csv("../Data/Processed/client_usage.csv", index=False)

print(df_client_usage.head())


#sampling to see how the clients who had multiple visits are mentioned
repeat_clients = (
    df_client_usage
        .groupby("client_id")["visits"]
        .sum()
        .loc[lambda x: x > 1]
        .index
)
df_client_usage[df_client_usage["client_id"].isin(repeat_clients)] \
    .sort_values(["client_id", "month"]) \
    .head(20)
