==> Notebooks/04_rq_experiment_metrics.ipynb <==
import pandas as pd
import os
import numpy as np

df_web = pd.read_csv("../Data/Processed/df_final_web_merged.csv")

## Experiment Metrics /  KPI Dataframe

df_web.columns.tolist()

df_web.head()

# Working with a copy to avoid data loss
df = df_web.copy()

# building one column for datetime
df["datetime"] = pd.to_datetime(df["date"].astype(str) + " " + df["time"].astype(str))

group_cols = ["client_id", "visitor_id", "visit_id"]

# sort events by visit date (and time)
df = df.sort_values(group_cols + ["datetime"])

# assigning step order for identifying backtracking
step_order_map = {
    "start": 0,
    "step_1": 1,
    "step_2": 2,
    "step_3": 3,
    "confirm": 4
}

df["step_order"] = df["process_step"].map(step_order_map)

# visit level summary for each client
summary = (
    df.groupby(group_cols)
      .agg(
          steps_count=("process_step", "count"),
          unique_steps=("process_step", "nunique"),
          start_time=("datetime", "min"),
          end_time=("datetime", "max")
      )
      .reset_index()
)

# completion (binary values)
completed_by_visit = (
    df.groupby(group_cols)["process_step"]
      .apply(lambda x: int("confirm" in x.values))
      .reset_index(name="completion_status")
)

summary = summary.merge(completed_by_visit, on=group_cols, how="left")

# drop-off step (if experiment not completed)
last_step = (
    df.groupby(group_cols)["process_step"]
      .last()
      .reset_index(name="last_step")
)

summary = summary.merge(last_step, on=group_cols, how="left")

summary["drop_off_step"] = np.where(
    summary["completion_status"] == 1,
    None,
    summary["last_step"]
)

summary.drop(columns="last_step", inplace=True)

# most frequent step (happens during backtracks)
most_freq = (
    df.groupby(group_cols)["process_step"]
      .agg(lambda x: x.value_counts().idxmax() if x.value_counts().max() > 1 else "none")
      .reset_index(name="most_frequent_step")
)

summary = summary.merge(most_freq, on=group_cols, how="left")

# total time and average time per step
summary["total_seconds"] = (summary["end_time"] - summary["start_time"]).dt.total_seconds()
summary["avg_time_per_step_sec"] = summary["total_seconds"] / summary["steps_count"]

summary["avg_time_per_step"] = summary["avg_time_per_step_sec"].apply(
    lambda x: f"{int(x//60)}:{int(x%60):02d}" if pd.notna(x) else None
)

# tracking backtracks
df["prev_step_order"] = df.groupby(group_cols)["step_order"].shift(1)
df["is_backtrack"] = ((df["step_order"] < df["prev_step_order"]) & df["step_order"].notna() & df["prev_step_order"].notna()).astype(int)

backtracks = (
    df.groupby(group_cols)["is_backtrack"]
      .agg(backtrack_count="sum")
      .reset_index()
)

backtracks["backtrack_flag"] = (backtracks["backtrack_count"] > 0).astype(int)

summary = summary.merge(backtracks, on=group_cols, how="left")


# Time to next step
df["next_datetime"] = df.groupby(group_cols)["datetime"].shift(-1)
df["delta_sec"] = (df["next_datetime"] - df["datetime"]).dt.total_seconds()

# Remove invalid time stamps
df.loc[df["delta_sec"] < 0, "delta_sec"] = np.nan

step_time = (
    df.groupby(group_cols + ["process_step"])["delta_sec"]
      .sum(min_count=1)
      .reset_index()
)

# Pivot - one column per step containing seconds spent in that particular step for the visit
step_time_wide = step_time.pivot_table(
    index=group_cols,
    columns="process_step",
    values="delta_sec",
    aggfunc="sum"
).reset_index()

# renaming columns
step_cols = [c for c in step_time_wide.columns if c not in group_cols]
step_time_wide = step_time_wide.rename(columns={c: f"time_in_{c}_sec" for c in step_cols})

summary = summary.merge(step_time_wide, on=group_cols, how="left")

# date / time fields
summary["date"] = summary["end_time"].dt.date

# Column order
df_summary = summary[[
    "client_id", "visitor_id", "visit_id",
    "steps_count", "unique_steps",
    "date", 
    "completion_status",          # binary 1/0
    "drop_off_step",
    "most_frequent_step",
    "avg_time_per_step",        
    "backtrack_flag",             # binary 1/0
    "backtrack_count",            # count of backward moves
    *[c for c in summary.columns if c.startswith("time_in_") and c.endswith("_sec")]  # time per step for all steps
]]


### Creating other total_time columns

df_summary = df_summary.copy()

# identify all per-step time columns
time_cols = [c for c in df_summary.columns if c.startswith("time_in_") and c.endswith("_sec")]

# total time in seconds
df_summary["total_time_sec"] = df_summary[time_cols].sum(axis=1, min_count=1)

# keep total time ONLY for completed visits
df_summary.loc[df_summary["completion_status"] != 1, "total_time_sec"] = np.nan

# time in mm:ss format
df_summary["total_time"] = df_summary["total_time_sec"].apply(
    lambda s: f"{int(s//60)}:{int(s%60):02d}" if pd.notna(s) else None
)

df_summary.to_csv(
    "../Data/Processed/df_web_visit_metrics_final.csv",
    index=False
)

df_summary.head()

----------------------------------------------------------------------------------------------------
----------------------------------------------------------------------------------------------------

## Merging Client Demographics and Experiment group info to KPI dataframe (Age, Gender and Variation)


# Load datasets with required info

df_exp = pd.read_csv("../Data/Processed/df_final_experiment_clients_clean.csv") #variation info dataframe
df_demo = pd.read_csv("../Data/Processed/df_final_demo_clean.csv") #demographics dataframe


# Selecting only wanted columns from datasets

df_exp = df_exp[["client_id", "variation"]]
df_demo = df_demo[["client_id", "clnt_age", "gender", "clnt_tenure_yr"]]


# Left merge to the visit metrics csv (df_summary)

df_summary = df_summary.merge(df_exp, on="client_id", how="left")
df_summary = df_summary.merge(df_demo, on="client_id", how="left")


# NaN values on variation and gender replaced by "excluded"
df_summary["variation"] = df_summary["variation"].fillna("excluded")
df_summary["gender"] = df_summary["gender"].fillna("Unknown")

# Client Age also has NaN values, replace with "unknown"

df_summary["clnt_age"] = df_summary["clnt_age"].fillna("Unknown")

df_summary.to_csv("../Data/Processed/df_web_visit_metrics_final.csv", index=False)


df_summary.head(20)

df_summary.shape

df_summary.columns

## Creating a DataFrame for only experiment participants


### Filter and create df using variation

df_exp_visits = df_summary[
    df_summary["variation"].isin(["Control", "Test"])
].copy() 

# this df is still visit level, with multiple rows per customer


df_exp_visits.head()

df_exp_visits.to_csv(
    "../Data/Processed/df_exp_visits.csv",
    index=False
)


### Create a client level df

# Dataframe with all experiment participants with one row per client

# client-level aggregation (1 row per client)
df_kpi_clients = (
    df_exp_visits
    .sort_values("date")
    .groupby("client_id", as_index=False)
    .agg({
        "variation": "first",                 # Control or Test
        "completion_status": "max",            # 1 if completed at least once
        "total_time_sec": "min",               # first / fastest successful completion
        "backtrack_flag": "max",               # if any backtracking
        "backtrack_count": "sum",              # total backtracks across multiple visits
        "steps_count": "sum",                  # total steps across multiple attempts
        "unique_steps": "max",                 # max unique steps seen
        "clnt_age": "first",
        "gender": "first",
        "clnt_tenure_yr": "first"
    })
)

# keep time only for clients who completed
df_kpi_clients.loc[df_kpi_clients["completion_status"] != 1, "total_time_sec"] = np.nan


df_kpi_clients.head()

df_kpi_clients.shape

- Shape matches the shape of the experiment_clients dataframe

# saving it as csv for later use
df_kpi_clients.to_csv(
    "../Data/Processed/df_kpi_clients.csv",
    index=False
)
